refactor: move small zero modules to lib/zero/
Part of #7697 — restructure src/lib/ into layered architecture.
Move model-provider, secret, variable, vm0-key, export, and
computer-connector to lib/zero/ as they manage Zero platform data.

Closes #7728
