+++ turbo/apps/web/app/api/integrations/github/route.ts
@@ -15,8 +15,8 @@ import {
   agentComposes,
   agentComposeVersions,
 } from "../../../../src/db/schema/agent-compose";
-import { listSecrets } from "../../../../src/lib/secret/secret-service";
-import { listVariables } from "../../../../src/lib/variable/variable-service";
+import { listSecrets } from "../../../../src/lib/zero/secret/secret-service";
+import { listVariables } from "../../../../src/lib/zero/variable/variable-service";
 import { listConnectors } from "../../../../src/lib/connector/connector-service";
 import type { AgentComposeYaml } from "../../../../src/types/agent-compose";
 import { resolveOrg } from "../../../../src/lib/org/resolve-org";
