+++ turbo/apps/web/app/api/zero/integrations/slack/route.ts
@@ -18,8 +18,8 @@ import {
   agentComposes,
   agentComposeVersions,
 } from "../../../../../src/db/schema/agent-compose";
-import { listSecrets } from "../../../../../src/lib/secret/secret-service";
-import { listVariables } from "../../../../../src/lib/variable/variable-service";
+import { listSecrets } from "../../../../../src/lib/zero/secret/secret-service";
+import { listVariables } from "../../../../../src/lib/zero/variable/variable-service";
 import { listConnectors } from "../../../../../src/lib/connector/connector-service";
 import { getOrgData } from "../../../../../src/lib/org/org-cache-service";
 import { createSlackClient } from "../../../../../src/lib/slack";
