+++ turbo/apps/web/app/api/zero/secrets/[name]/route.ts
@@ -10,7 +10,7 @@ import {
   isAuthError,
 } from "../../../../../src/lib/auth/require-auth";
 import { resolveOrg } from "../../../../../src/lib/org/resolve-org";
-import { deleteSecret } from "../../../../../src/lib/secret/secret-service";
+import { deleteSecret } from "../../../../../src/lib/zero/secret/secret-service";
 import { logger } from "../../../../../src/lib/logger";
 import { isNotFound } from "../../../../../src/lib/errors";
