+++ turbo/apps/web/src/lib/zero/__tests__/build-zero-context.test.ts
@@ -13,9 +13,9 @@ import {
   findTestRunnerJobEntry,
 } from "../../../__tests__/api-test-helpers";
 import { createZeroRun } from "../zero-run-service";
-import { upsertOrgModelProvider } from "../../model-provider/model-provider-service";
-import { upsertSecretByOrg } from "../../secret/secret-service";
-import { setVariable } from "../../variable/variable-service";
+import { upsertOrgModelProvider } from "../model-provider/model-provider-service";
+import { upsertSecretByOrg } from "../secret/secret-service";
+import { setVariable } from "../variable/variable-service";
 import { ORG_SENTINEL_USER_ID } from "../../org/org-sentinel";
 import { isNoModelProvider } from "../../errors";
 import { reloadEnv } from "../../../env";
