+++ turbo/apps/web/src/lib/zero/computer-connector/computer-connector-service.ts
@@ -7,11 +7,11 @@
 import { randomUUID } from "crypto";
 import { eq, and } from "drizzle-orm";
 import type { ComputerConnectorCreateResponse } from "@vm0/core";
-import { connectors } from "../../db/schema/connector";
-import { secrets } from "../../db/schema/secret";
-import { decryptSecretValue } from "../shared/crypto";
-import { badRequest, conflict, notFound } from "../errors";
-import { logger } from "../logger";
+import { connectors } from "../../../db/schema/connector";
+import { secrets } from "../../../db/schema/secret";
+import { decryptSecretValue } from "../../shared/crypto";
+import { badRequest, conflict, notFound } from "../../errors";
+import { logger } from "../../logger";
 import { upsertSecretByOrg } from "../secret/secret-service";
 import {
   findOrCreateBotUser,
