+++ turbo/apps/web/src/lib/zero/model-provider/__tests__/org-model-provider.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, vi } from "vitest";
-import { testContext } from "../../../__tests__/test-helpers";
+import { testContext } from "../../../../__tests__/test-helpers";
 import {
   listOrgModelProviders,
   upsertOrgModelProvider,
