+++ turbo/apps/web/src/lib/zero/model-provider/model-provider-service.ts
@@ -10,12 +10,12 @@ import {
   type ModelProviderType,
   type ModelProviderFramework,
 } from "@vm0/core";
-import { modelProviders } from "../../db/schema/model-provider";
-import { secrets } from "../../db/schema/secret";
-import { encryptSecretValue } from "../shared/crypto";
-import { badRequest, notFound } from "../errors";
-import { logger } from "../logger";
-import { ORG_SENTINEL_USER_ID } from "../org/org-sentinel";
+import { modelProviders } from "../../../db/schema/model-provider";
+import { secrets } from "../../../db/schema/secret";
+import { encryptSecretValue } from "../../shared/crypto";
+import { badRequest, notFound } from "../../errors";
+import { logger } from "../../logger";
+import { ORG_SENTINEL_USER_ID } from "../../org/org-sentinel";
 
 const log = logger("service:model-provider");
