+++ turbo/apps/web/src/lib/zero/secret/__tests__/org-secret-service.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, vi } from "vitest";
-import { testContext } from "../../../__tests__/test-helpers";
+import { testContext } from "../../../../__tests__/test-helpers";
 import {
   listOrgSecrets,
   setOrgSecret,
