+++ turbo/apps/web/src/lib/zero/secret/secret-service.ts
@@ -1,10 +1,10 @@
 import { eq, and } from "drizzle-orm";
 import { type SecretType } from "@vm0/core";
-import { secrets } from "../../db/schema/secret";
-import { encryptSecretValue, decryptSecretValue } from "../shared/crypto";
-import { badRequest, notFound } from "../errors";
-import { logger } from "../logger";
-import { ORG_SENTINEL_USER_ID } from "../org/org-sentinel";
+import { secrets } from "../../../db/schema/secret";
+import { encryptSecretValue, decryptSecretValue } from "../../shared/crypto";
+import { badRequest, notFound } from "../../errors";
+import { logger } from "../../logger";
+import { ORG_SENTINEL_USER_ID } from "../../org/org-sentinel";
 
 const log = logger("service:secret");
